discard one decimal digit on scores when they pass 2000
A Hardcore AI team in Zen speed Team Race gets less than 100 points per lap, so I assume the score of a team won't pass 10000 in a 99-lap Team Race.

And a width that can fit 2000 in the hud font can fit up to 19999.

For Eliminator, I am just putting it there as a consistency with the team score panel, as the max target score the game allows is less than 1000.