Update README.md
The camera of player 1 was stretched to fill the screen during the whole development time, so I thought it's impossible to make a custom hud to support the mode. The camera was "fixed" around the time I released the hud, but it was too late for me to spend more resource on making the hud properly working.