extend the shake duration when the current shake amount is bigger
shifter was doing nothing when the current amount is bigger than the amount it is trying to apply, making a weird situation where the shake trigger is caused at the near end of an ongoing shake duration. It might look like the last mine hit in the line "stabilized" the hud.

Not sure if it's okay to not change the amount, but I came blank on what to do with it, so I am going to leave it as it is. At least, the duration given is fixed at a short time of 0.25s.